+++ src/jk_utils/datetime/TApprox.py
@@ -11,7 +11,7 @@ import datetime
 #
 # This class represents a time value with a 1 minute accuracy.
 #
-class T1(object):
+class TApprox(object):
 
 	################################################################################################################################
 	## Constructor
@@ -21,11 +21,17 @@ class T1(object):
 		self._hour = 0
 		self._minute = 0
 		self._absoluteTick = 0
+		self._isApproximate = False
 	#
 
 	################################################################################################################################
 	## Public Properties
 	################################################################################################################################
+
+	@property
+	def isApproximate(self) -> bool:
+		return self._isApproximate
+	#
 
 	@property
 	def hour(self) -> int:
@@ -52,7 +58,7 @@ class T1(object):
 	################################################################################################################################
 
 	def clone(self):
-		ret = T1()
+		ret = TApprox()
 		ret._hour = self._hour
 		ret._minute = self._minute
 		ret._absoluteTick = self._absoluteTick
@@ -64,51 +70,57 @@ class T1(object):
 	################################################################################################################################
 
 	def __str__(self):
-		return "{:02d}:{:02d}".format(self._hour, self._minute)
+		if self._isApproximate:
+			return "ca. {:02d}:{:02d}".format(self._hour, self._minute)
+		else:
+			return "{:02d}:{:02d}".format(self._hour, self._minute)
 	#
 
 	def __repr__(self):
-		return "{:02d}:{:02d}".format(self._hour, self._minute)
+		if self._isApproximate:
+			return "ca. {:02d}:{:02d}".format(self._hour, self._minute)
+		else:
+			return "{:02d}:{:02d}".format(self._hour, self._minute)
 	#
 
 	def __eq__(self, other):
-		if isinstance(other, T1):
+		if isinstance(other, TApprox):
 			return self._absoluteTick == other._absoluteTick
 		else:
 			return False
 	#
 
 	def __ne__(self, other):
-		if isinstance(other, T1):
+		if isinstance(other, TApprox):
 			return self._absoluteTick != other._absoluteTick
 		else:
 			return True
 	#
 
 	def __ge__(self, other):
-		assert isinstance(other, T1)
+		assert isinstance(other, TApprox)
 		return self._absoluteTick >= other._absoluteTick
 	#
 
 	def __gt__(self, other):
-		assert isinstance(other, T1)
+		assert isinstance(other, TApprox)
 		return self._absoluteTick > other._absoluteTick
 	#
 
 	def __le__(self, other):
-		assert isinstance(other, T1)
+		assert isinstance(other, TApprox)
 		return self._absoluteTick <= other._absoluteTick
 	#
 
 	def __lt__(self, other):
-		assert isinstance(other, T1)
+		assert isinstance(other, TApprox)
 		return self._absoluteTick < other._absoluteTick
 	#
 
 	def __iadd__(self, other):
 		if isinstance(other, int):
 			self._absoluteTick += other
-		elif isinstance(other, T1):
+		elif isinstance(other, TApprox):
 			self._absoluteTick += other._absoluteTick
 		else:
 			raise Exception(str(type(other)))
@@ -120,7 +132,7 @@ class T1(object):
 	def __isub__(self, other):
 		if isinstance(other, int):
 			self._absoluteTick -= other
-		elif isinstance(other, T1):
+		elif isinstance(other, TApprox):
 			self._absoluteTick -= other._absoluteTick
 		else:
 			raise Exception(str(type(other)))
@@ -131,18 +143,18 @@ class T1(object):
 
 	def __add__(self, other):
 		if isinstance(other, int):
-			return T1.createFromAbsoluteTick(self._absoluteTick + other)
-		elif isinstance(other, T1):
-			return T1.createFromAbsoluteTick(self._absoluteTick + other._absoluteTick)
+			return TApprox.createFromAbsoluteTick(self._absoluteTick + other)
+		elif isinstance(other, TApprox):
+			return TApprox.createFromAbsoluteTick(self._absoluteTick + other._absoluteTick)
 		else:
 			raise Exception(str(type(other)))
 	#
 
 	def __sub__(self, other):
 		if isinstance(other, int):
-			return T1.createFromAbsoluteTick(self._absoluteTick - other)
-		elif isinstance(other, T1):
-			return T1.createFromAbsoluteTick(self._absoluteTick - other._absoluteTick)
+			return TApprox.createFromAbsoluteTick(self._absoluteTick - other)
+		elif isinstance(other, TApprox):
+			return TApprox.createFromAbsoluteTick(self._absoluteTick - other._absoluteTick)
 		else:
 			raise Exception(str(type(other)))
 	#
@@ -157,9 +169,11 @@ class T1(object):
 
 	@staticmethod
 	def __tryParse(s:str):
-		for patternType, pattern in [
-			("hms", r"(?P<hour>\d\d):(?P<minute>\d\d):(?P<second>\d\d)"),
-			("hm", r"(?P<hour>\d\d):(?P<minute>\d\d)"),
+		for approx, patternType, pattern in [
+			(False, "hms", r"^(?P<hour>\d\d):(?P<minute>\d\d):(?P<second>\d\d)"),
+			(False,  "hm", r"^(?P<hour>\d\d):(?P<minute>\d\d)"),
+			(True, "hms", r"^ca\.?\s*(?P<hour>\d\d):(?P<minute>\d\d):(?P<second>\d\d)"),
+			(True,  "hm", r"^ca\.?\s*(?P<hour>\d\d):(?P<minute>\d\d)"),
 		]:
 			m = re.match("^" + pattern + "$", s)
 			if m:
@@ -167,58 +181,79 @@ class T1(object):
 					hour = int(m.group("hour"))
 					minute = int(m.group("minute"))
 					second = int(m.group("second"))
+					posR = m.end()
 				elif patternType == "hm":
 					hour = int(m.group("hour"))
 					minute = int(m.group("minute"))
 					second = 0
+					posR = m.end()
 				else:
 					raise Exception()
 				if (0 <= hour <= 23) and (0 <= minute <= 59) and (0 <= second <= 59):
-					yield (hour, minute, second)
+					yield (approx, hour, minute, second, posR)
 	#
 
+	#
+	# Tries to parse a text string such as:
+	# * "23:59:00"
+	# * "23:59"
+	# * "ca. 23:59:00"
+	# * "ca. 23:59"
+	# * "ca.23:59:00"
+	# * "ca.23:59"
+	# * "ca 23:59:00"
+	# * "ca 23:59"
+	#
+	# @return		TApprox		The parsed date or <c>None</c> if unparsable
+	# @return		int			The number of characters parsed (= <c>0</c> if unparsable)
+	#
 	@staticmethod
-	def tryParseFromStr(s:str):
+	def tryParseFromStr(s:str) -> tuple:
 		if isinstance(s, str):
 			s = s.strip()
-			for hour, minute, second in T1.__tryParse(s):
-				return T1.createFromTime(hour, minute)
-		return None
+			for approx, hour, minute, second, posR in TApprox.__tryParse(s):
+				return TApprox.createFromTime(hour, minute, approx), posR
+		return None, 0
 	#
 
 	@staticmethod
-	def createFromTime(hour:int, minute:int):
+	def createFromTime(hour:int, minute:int, isApproximate:bool = False):
 		assert isinstance(hour, int)
 		assert isinstance(minute, int)
+		assert isinstance(isApproximate, bool)
 
-		ret = T1()
+		ret = TApprox()
 		ret._hour = hour
 		ret._minute = minute
 		ret._absoluteTick = ret._hour * 60 + ret._minute
+		ret._isApproximate = isApproximate
 		return ret
 	#
 
 	@staticmethod
-	def createFromAbsoluteTick(absoluteTick:int):
+	def createFromAbsoluteTick(absoluteTick:int, isApproximate:bool = False):
 		assert isinstance(absoluteTick, int)
 
-		ret = T1()
+		ret = TApprox()
 		ret._hour = absoluteTick // 60
 		ret._minute = absoluteTick % 60
 		ret._absoluteTick = absoluteTick
+		ret._isApproximate = isApproximate
 		return ret
 	#
 
 	@staticmethod
-	def createFromHourMinutes(hourMinutes:int):
+	def createFromHourMinutes(hourMinutes:int, isApproximate:bool = False):
 		assert isinstance(hourMinutes, int)
+		assert isinstance(isApproximate, bool)
 
-		ret = T1()
+		ret = TApprox()
 		ret._hour = hourMinutes // 100
 		ret._minute = hourMinutes % 100
 		if (ret._hour < 0) or (ret._hour > 23) or (ret._minute < 0) or (ret._minute > 59):
 			raise Exception("Invalid time!")
 		ret._absoluteTick = ret._hour * 60 + ret._minute
+		ret._isApproximate = isApproximate
 		return ret
 	#
 
@@ -226,7 +261,7 @@ class T1(object):
 	def now():
 		dt = datetime.datetime.now()
 
-		ret = T1()
+		ret = TApprox()
 		ret._hour = dt.hour
 		ret._minute = dt.minute
 		ret._absoluteTick = ret._hour * 60 + ret._minute