+++ src/jk_utils/datetime/__init__.py
@@ -1,12 +1,13 @@
 ﻿
 
 
-__version__ = "0.2021.2.16"
+__version__ = "0.2021.6.29"
 
 
 from .D import D
 from .T1 import T1
 from .T5 import T5
+from .TApprox import TApprox
 
 from .dateiterator import dateRange, dateRangeI
 from .timeiterator5 import timeRange5